# Conflicts:
#	submarine/src/main/java/com/AEProjekt/submarine/figures/Submarine.java
#	submarine/src/main/java/com/AEProjekt/submarine/levelz/Level3.java
#	submarine/src/main/java/com/AEProjekt/submarine/levelz/Level4.java
#	submarine/src/main/java/com/AEProjekt/submarine/users/User.java

+++ submarine/src/test/java/com/AEProjekt/submarine/InputLevel/TestStructureForInputLevel3.java
@@ -2,5 +2,5 @@ package com.AEProjekt.submarine.InputLevel;
 
 public class TestStructureForInputLevel3 {
     public boolean right_or_wrong;
-    public double pX;
+    public int pX;
 }
